Add logic to handle open unit test results
