+++ apps/vs-code-designer/src/app/commands/workflows/unitTest/openUnitTestResults.ts
@@ -2,15 +2,38 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import { getUnitTestName } from '../../../utils/unitTests';
-import { getWorkflowNode } from '../../../utils/workspace';
+import { developmentDirectoryName, testsDirectoryName, workflowFileName } from '../../../../constants';
+import { getUnitTestName, pickUnitTest } from '../../../utils/unitTests';
+import { tryGetLogicAppProjectRoot } from '../../../utils/verifyIsProject';
+import { getWorkflowNode, getWorkspaceFolder } from '../../../utils/workspace';
 import { type IAzureConnectorsContext } from '../azureConnectorWizard';
 import OpenDesignerForLocalProject from '../openDesigner/openDesignerForLocalProject';
-import type * as vscode from 'vscode';
+import { readFileSync } from 'fs';
+import * as path from 'path';
+import { type TestItem, Uri } from 'vscode';
 
-export async function openUnitTestResults(context: IAzureConnectorsContext, node: vscode.Uri): Promise<void> {
-  const unitTestName = getUnitTestName(node.fsPath);
-  const workflowNode = getWorkflowNode(node) as vscode.Uri;
-  const openDesignerObj = new OpenDesignerForLocalProject(context, workflowNode, unitTestName);
+export async function openUnitTestResults(context: IAzureConnectorsContext, node: Uri | TestItem): Promise<void> {
+  let unitTestNode: Uri;
+  const workspaceFolder = await getWorkspaceFolder(context);
+  const projectPath = await tryGetLogicAppProjectRoot(context, workspaceFolder);
+
+  if (node && node instanceof Uri) {
+    unitTestNode = getWorkflowNode(node) as Uri;
+  } else if (node && !(node instanceof Uri) && node.uri instanceof Uri) {
+    unitTestNode = node.uri;
+  } else {
+    const unitTest = await pickUnitTest(context, path.join(projectPath, developmentDirectoryName, testsDirectoryName));
+    unitTestNode = Uri.file(unitTest.data) as Uri;
+  }
+
+  const workflowName = path.basename(path.dirname(unitTestNode.fsPath));
+  const workflowPath = path.join(projectPath, workflowName, workflowFileName);
+  const workflowNode = Uri.file(workflowPath);
+  const unitTestDefinition = JSON.parse(readFileSync(unitTestNode.fsPath, 'utf8'));
+  const unitTestName = getUnitTestName(unitTestNode.fsPath);
+
+  const openDesignerObj = new OpenDesignerForLocalProject(context, workflowNode, unitTestName, unitTestDefinition);
+  await openDesignerObj?.createPanel();
+
   await openDesignerObj?.createPanel();
 }
